+++ src/mocks/browser.ts
@@ -1,5 +1,16 @@
 import { setupWorker } from 'msw';
-import { handlers } from './handlers';
+
+import {
+  productHandlers,
+  cartProductHandlers,
+  couponHandlers,
+  orderHandlers,
+} from './handlers';
 
 // Configure the Service Worker for in-browser request interception
-export const worker = setupWorker(...handlers);
+export const worker = setupWorker(
+  ...productHandlers,
+  ...cartProductHandlers,
+  ...couponHandlers,
+  ...orderHandlers
+);
