+++ src/mocks/handlers/index.ts
@@ -1,0 +1,4 @@
+export * from './productHandlers';
+export * from './cartProductHandlers';
+export * from './orderHandlers';
+export * from './couponHandlers';
